Revamp configs
