+++ packages/synapse-constants/rollup.config.js
@@ -6,45 +6,48 @@ import terser from '@rollup/plugin-terser'
 import url from '@rollup/plugin-url'
 import { codecovRollupPlugin } from '@codecov/rollup-plugin'
 
-export default {
-  input: 'src/index.ts', // Entry point for your constants/utilities
-  output: [
-    {
-      file: 'dist/bundle.cjs.js',
-      format: 'cjs', // CommonJS output
-      sourcemap: true,
-    },
-    {
-      file: 'dist/bundle.esm.js',
-      format: 'esm', // ES Module output
-      sourcemap: true,
-    },
-  ],
-  plugins: [
-    nodeResolve({
-      preferBuiltins: true,
-    }),
-    commonjs(),
-    json(),
-    typescript({
-      tsconfig: './tsconfig.json',
-      declaration: true,
-      declarationDir: 'dist',
-    }),
-    terser(),
-    url({
-      include: ['**/*.svg', '**/*.png', '**/*.jpg'],
-      limit: 0,
-      emitFiles: false,
-    }),
-    codecovRollupPlugin({
-      enableBundleAnalysis: process.env.CODECOV_TOKEN !== undefined,
-      bundleName: 'synapse-constants',
-      uploadToken: process.env.CODECOV_TOKEN,
-      uploadOverrides: {
-        sha: process.env.GH_COMMIT_SHA,
+import packageJson from './package.json'
+
+export default [
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        file: packageJson.main,
+        format: 'cjs',
       },
-    }),
-  ],
-  external: ['lodash', 'ethers'],
-}
+      {
+        file: packageJson.module,
+        format: 'esm',
+      },
+    ],
+    plugins: [
+      nodeResolve({
+        preferBuiltins: true,
+      }),
+      commonjs(),
+      json(),
+      typescript({
+        tsconfig: './tsconfig.json',
+        declaration: true,
+        declarationDir: './dist/types',
+        useTsconfigDeclarationDir: true,
+      }),
+      terser(),
+      url({
+        include: ['**/*.svg', '**/*.png', '**/*.jpg'],
+        limit: 20000,
+        emitFiles: false,
+      }),
+      codecovRollupPlugin({
+        enableBundleAnalysis: process.env.CODECOV_TOKEN !== undefined,
+        bundleName: 'synapse-constants',
+        uploadToken: process.env.CODECOV_TOKEN,
+        uploadOverrides: {
+          sha: process.env.GH_COMMIT_SHA,
+        },
+      }),
+    ],
+    external: ['lodash', 'ethers'],
+  },
+]
