+++ packages/synapse-constants/src/constants/index.ts
@@ -1,4 +1,4 @@
 export * as TOKENS from './tokens'
 export * as CHAINS from './chains'
-export * from './types'
+export * from '../types'
 export * from './assets'
