+++ packages/synapse-constants/src/constants/tokens/auxilliary.ts
@@ -4,7 +4,7 @@ import ethLogo from '../assets/icons/eth.svg'
 import mimLogo from '../assets/icons/mim.svg'
 import usdcLogo from '../assets/icons/usdc.svg'
 import usdtLogo from '../assets/icons/usdt.svg'
-import { Token } from '../types'
+import { Token } from '../../types'
 import * as CHAINS from '../chains/master'
 
 export const AVWETH = new Token({
