+++ packages/synapse-constants/src/constants/tokens/poolMaster.ts
@@ -1,5 +1,5 @@
 import synapseLogo from '../assets/icons/syn.svg'
-import { Token } from '../types'
+import { Token } from '../../types'
 import * as CHAINS from '../chains/master'
 import {
   BUSD,
