+++ packages/synapse-constants/src/constants/tokens/swapMaster.ts
@@ -1,4 +1,4 @@
-import { Token } from '../types'
+import { Token } from '../../types'
 import usdtLogo from '../assets/icons/usdt.svg'
 import usdcLogo from '../assets/icons/usdc.svg'
 import * as CHAINS from '../chains/master'
