+++ packages/synapse-constants/src/index.ts
@@ -1,6 +1,9 @@
 export * from './constants/tokens/index'
 export * as CHAINS from './constants/chains/index'
-export * from './constants/types/index'
-// export * from './constants/assets/chains/index'
-// export * from './constants/assets/explorer/index'
-// export * from './constants/assets/icons/index'
+export * from './types/index'
+export * from './constants/assets/chains/index'
+export * from './constants/assets/explorer/index'
+export * from './constants/assets/icons/index'
+
+export { MINICHEF_ADDRESSES } from './constants/minichef'
+export { BRIDGE_MAP } from './constants/bridgeMap'
