+++ packages/synapse-constants/src/utils/validateAddresses.ts
@@ -1,0 +1,13 @@
+import { getAddress } from '@ethersproject/address'
+
+export const validateAddresses = (addresses: {
+  [x: number]: string
+}): { [x: number]: string } => {
+  const reformatted: { [x: number]: string } = {}
+  for (const chainId in addresses) {
+    reformatted[chainId] = addresses[chainId]
+      ? getAddress(addresses[chainId])
+      : ''
+  }
+  return reformatted
+}
